requirements.txt
# APT Fitness Assistant - Complete Dependencies
# Updated for restructured modular architecture

# Core web framework
streamlit>=1.29.0
streamlit-option-menu>=0.3.0

# Computer vision and AI
opencv-python>=4.9.0.80
mediapipe>=0.10.9
numpy>=1.26.2
pillow>=10.1.0

# Data analysis and visualization
pandas>=2.1.4
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
scikit-learn>=1.3.2

# Deep learning
tensorflow>=2.15.0,<2.17.0
keras>=2.15.0
scipy>=1.10.0

# Image processing
albumentations>=1.3.0
imagehash>=4.3.0
opencv-contrib-python>=4.8.0
imageio>=2.31.0
scikit-image>=0.21.0

# Database and persistence
# sqlite3 is included in Python standard library

# Web API (optional for future API endpoints)
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
werkzeug>=2.3.0

# Utilities
python-dateutil>=2.8.2
typing-extensions>=4.0.0
pathlib2>=2.3.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
